changelog: -added a dedicated locations command that will list all of our available locations -finally got the bot to respond to message via DMs and assign the approriate role -changed $numofusers to a student counter ($students).
I started the foundations for locking out the other roles so that students don't assign themselves to multiples locations but so far only kinloch is functioning properly, NMB is somewhat functional, and nautilus is borked.
I commented out the functions that would assign roles within the discord since the dm functions also work within the discord server as well.
This will get changed in future versions so that roles can only be assigned via DMs.