changelog: -added staff, student, and password commands (the first two being role assignments) -added a help command (commands) that will link to the wiki on github for a complete list o> -cleaned up $students by removing some unnecessary lines
I added some more commands to Reginald that I think will be useful.
The password one is one that I think will come in handy when we decide to start rolling out machine learning
There are still a couple more commands that I want to add, but other than those. This will probably be the end of the commands branch.
From here, I'll probably focus on removing the indexes for the commands that deal with roles and instead of looking for a specific role in an index, I'll have it search for a role with the name instead.
I need to start planning that out, but once I finished all the commands that I think I'll implement for Reginald 1.0
I also need to work on the wiki for Reginald.